Power Card Population now separate file
