+++ api/ExpressedRealms.Powers.Reporting/powerCards/CardTypes/PowerCards/PowerCardData.cs
